Bug: buttons display property now changes properly
When moving to Sass I added an auxiliary class, .flex-center, at the
bottom of the stylesheet. This class was added to the buttons id after
the media queries were applied, causing the `display: none` property to
be replaed by `display: flex`. This was fixed by moving the media
queries to the bottom of the file.